requirements.txt
paho_mqtt==1.6.1
pyowm==3.3.0
